requirements.txt
psycopg2-binary==2.9.3
SQLAlchemy==1.3.24
SQLAlchemy-Utils==0.37.9
boto3==1.24.64
pandas==1.4.4
python-decouple==3.6
numpy==1.23.2